+++ CHANGELOG.md
@@ -3,7 +3,11 @@
 ## Unreleased
 
 ### Added
+- The Sender Class and Event Name webhook settings now show suggestions based on the available classes and events.
 - Webhooks can now have custom payloads. ([#3](https://github.com/craftcms/webhooks/pull/3))
+
+### Changed
+- Webhooks now requires Craft 3.1 or later.
 
 ## 1.1.2 - 2018-12-21